marketOpen/Close機能とversion表示追加

+++ src/main/java/red/man10/VaultManager.java
@@ -52,7 +52,7 @@ public class VaultManager {
     public void showBalance(UUID uuid){
         OfflinePlayer p = Bukkit.getOfflinePlayer(uuid).getPlayer();
         double money = getBalance(uuid);
-        p.getPlayer().sendMessage(ChatColor.YELLOW + "あなたの所持金は$" + (int)money);
+        p.getPlayer().sendMessage(ChatColor.YELLOW + "あなたの所持金は$" + money);
     }
     /////////////////////////////////////
     //      引き出し
@@ -67,7 +67,7 @@ public class VaultManager {
         if(resp.transactionSuccess()){
             if(p.isOnline()) {
                 if(showMessage) {
-                    p.getPlayer().sendMessage(ChatColor.YELLOW + "$" + (int) money + "支払いました");
+                    p.getPlayer().sendMessage(ChatColor.YELLOW + "$" +  money + "支払いました");
                 }
             }
             return true;
@@ -88,7 +88,7 @@ public class VaultManager {
         if(resp.transactionSuccess()){
             if(p.isOnline()){
                 if(showMessage){
-                    p.getPlayer().sendMessage(ChatColor.YELLOW + "$"+(int)money+"受取りました");
+                    p.getPlayer().sendMessage(ChatColor.YELLOW + "$"+ money+"受取りました");
 
                 }
             }
